Reorganizing some code.
