+++ src/uvcsite/plugins/configure.zcml
@@ -1,0 +1,18 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:browser="http://namespaces.zope.org/browser"
+    xmlns:grok="http://namespaces.zope.org/grok">
+
+  <browser:defaultView
+      for=".components.IPlugin"
+      name="index"
+      />
+
+  <browser:defaultView
+      for=".panel.PluginsPanel"
+      name="index"
+      />
+
+   <grok:grok package="." />
+
+</configure>
